+++ CesiumGltfReader/src/decodeDraco.cpp
@@ -126,10 +126,10 @@ void copyDecodedIndices(
     return;
   }
 
-  if (pIndicesAccessor->count != pMesh->num_faces() * 3) {
+  if (pIndicesAccessor->count != int64_t(pMesh->num_faces() * 3)) {
     readGltf.warnings.emplace_back(
         "indices accessor doesn't match with decoded Draco indices");
-    pIndicesAccessor->count = pMesh->num_faces() * 3;
+    pIndicesAccessor->count = int64_t(pMesh->num_faces() * 3);
   }
 
   draco::PointIndex::ValueType numPoint = pMesh->num_points();
@@ -237,7 +237,7 @@ void copyDecodedAttribute(
 
   const int8_t numberOfComponents = pAccessor->computeNumberOfComponents();
   const int64_t stride =
-      numberOfComponents * pAccessor->computeByteSizeOfComponent();
+      int64_t(numberOfComponents * pAccessor->computeByteSizeOfComponent());
   const int64_t sizeBytes = pAccessor->count * stride;
 
   buffer.cesium.data.resize(static_cast<size_t>(sizeBytes));
