+++ CesiumQuantizedMeshTerrain/test/TestQuantizedMeshContent.cpp
@@ -418,7 +418,7 @@ void checkGridMesh(
   int32_t v = 0;
 
   std::vector<glm::dvec2> uvs;
-  uvs.reserve(verticesWidth * verticesHeight);
+  uvs.reserve(size_t(verticesWidth * verticesHeight));
   uint32_t positionIdx = 0;
   uint32_t idx = 0;
   for (uint32_t y = 0; y < verticesHeight; ++y) {
@@ -490,8 +490,9 @@ void checkGridMesh(
   size_t eastIndicesCount = quantizedMesh.vertexData.eastIndices.size();
   size_t northIndicesCount = quantizedMesh.vertexData.northIndices.size();
 
-  size_t gridVerticesCount = verticesWidth * verticesHeight;
-  size_t gridIndicesCount = (verticesHeight - 1) * (verticesWidth - 1) * 6;
+  size_t gridVerticesCount = size_t(verticesWidth * verticesHeight);
+  size_t gridIndicesCount =
+      size_t((verticesHeight - 1) * (verticesWidth - 1) * 6);
   size_t totalSkirtVertices = westIndicesCount + southIndicesCount +
                               eastIndicesCount + northIndicesCount;
   size_t totalSkirtIndices = (totalSkirtVertices - 4) * 6;
@@ -592,7 +593,8 @@ static void checkGeneratedGridNormal(
     uint32_t verticesWidth,
     uint32_t verticesHeight) {
   uint32_t totalGridIndices = (verticesWidth - 1) * (verticesHeight - 1) * 6;
-  std::vector<glm::vec3> expectedNormals(verticesWidth * verticesHeight);
+  std::vector<glm::vec3> expectedNormals(
+      size_t(verticesWidth * verticesHeight));
   for (uint32_t i = 0; i < totalGridIndices; i += 3) {
     I id0 = indices[i];
     I id1 = indices[i + 1];
@@ -638,7 +640,7 @@ static void checkGeneratedGridNormal(
   size_t eastIndicesCount = quantizedMesh.vertexData.eastIndices.size();
   size_t northIndicesCount = quantizedMesh.vertexData.northIndices.size();
 
-  size_t gridVerticesCount = verticesWidth * verticesHeight;
+  size_t gridVerticesCount = size_t(verticesWidth * verticesHeight);
   size_t totalSkirtVertices = westIndicesCount + southIndicesCount +
                               eastIndicesCount + northIndicesCount;
 
@@ -1025,7 +1027,8 @@ TEST_CASE("Test converting quantized mesh to gltf with skirt") {
     glm::vec3 normal = glm::normalize(glm::vec3(0.2, 1.4, 0.3));
     uint8_t x = 0, y = 0;
     octEncode(normal, x, y);
-    std::vector<std::byte> octNormals(verticesWidth * verticesHeight * 2);
+    std::vector<std::byte> octNormals(
+        size_t(verticesWidth * verticesHeight * 2));
     for (size_t i = 0; i < octNormals.size(); i += 2) {
       octNormals[i] = std::byte(x);
       octNormals[i + 1] = std::byte(y);
@@ -1067,7 +1070,7 @@ TEST_CASE("Test converting quantized mesh to gltf with skirt") {
 
     REQUIRE(
         static_cast<size_t>(normals.size()) ==
-        (verticesWidth * verticesHeight + totalSkirtVerticesCount));
+        size_t(verticesWidth * verticesHeight) + totalSkirtVerticesCount);
     for (int64_t i = 0; i < normals.size(); ++i) {
       REQUIRE(Math::equalsEpsilon(normals[i].x, normal.x, Math::Epsilon2));
       REQUIRE(Math::equalsEpsilon(normals[i].y, normal.y, Math::Epsilon2));
@@ -1112,7 +1115,7 @@ TEST_CASE("Test converting ill-formed quantized mesh") {
   glm::vec3 normal = glm::normalize(glm::vec3(0.2, 1.4, 0.3));
   uint8_t x = 0, y = 0;
   octEncode(normal, x, y);
-  std::vector<std::byte> octNormals(verticesWidth * verticesHeight * 2);
+  std::vector<std::byte> octNormals(size_t(verticesWidth * verticesHeight * 2));
   for (size_t i = 0; i < octNormals.size(); i += 2) {
     octNormals[i] = std::byte(x);
     octNormals[i + 1] = std::byte(y);
